my-agent/requirements.txt
# Core dependencies
strands-agents>=1.12.0
strands-agents-tools>=0.2.11
boto3>=1.34.0
botocore>=1.34.0

# Data handling
pydantic>=2.0.0
python-dateutil>=2.8.0

# Utilities
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
